requirements.txt
asgiref==3.5.2
astroid==2.12.9
autopep8==1.7.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.5
dill==0.3.5.1
Django==4.1.3
django-autoslug-field==0.2.3
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.4.0
platformdirs==2.5.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyClamd==0.4.0
pycodestyle==2.9.1
pylint==2.15.2
pyparsing==3.0.9
python-magic==0.4.27
pytz==2022.6
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.14.1
